edit

+++ printf/ft_utils_second.c
@@ -6,7 +6,7 @@
 /*   By: hkubo <hkubo@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/21 22:05:33 by hkubo             #+#    #+#             */
-/*   Updated: 2021/03/25 11:24:51 by hkubo            ###   ########.fr       */
+/*   Updated: 2021/03/25 11:39:04 by hkubo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ int		field_no_precision_sub(long num, char *str, t_plist flag_list, int i)
 	if (num < 0 && flag_list.flag[0] != 1 && flag_list.flag[1] == 1)
 	{
 		write(1, "-", 1);
-		while ((flag_list.field - ft_strlen(str) - (i++)) > 0)
+		while ((flag_list.field - (int)ft_strlen(str) - (i++)) > 0)
 			write(1, "0", 1);
 		num *= -1;
 		if (!(tmp = ft_itoa(num)))
@@ -49,7 +49,7 @@ int		field_no_precision_sub(long num, char *str, t_plist flag_list, int i)
 	}
 	else
 	{
-		while ((flag_list.field - ft_strlen(str) - (i++)) > 0)
+		while ((flag_list.field - (int)ft_strlen(str) - (i++)) > 0)
 		{
 			if (flag_list.flag[1] == 1)
 				write(1, "0", 1);
@@ -72,7 +72,7 @@ char	*field_precision_sub(int num, char *str, t_plist flag_list, int i)
 			return (NULL);
 		tmp[0] = '-';
 		i = 1;
-		while ((flag_list.precision + 2 - ft_strlen(str) - i) > 0)
+		while ((flag_list.precision + 2 - (int)ft_strlen(str) - i) > 0)
 			tmp[i++] = '0';
 		j = 1;
 	}
@@ -80,7 +80,7 @@ char	*field_precision_sub(int num, char *str, t_plist flag_list, int i)
 	{
 		if (!(tmp = (char *)malloc(sizeof(char) * (flag_list.precision + 1))))
 			return (NULL);
-		while ((flag_list.precision - ft_strlen(str) - i) > 0)
+		while ((flag_list.precision - (int)ft_strlen(str) - i) > 0)
 			tmp[i++] = '0';
 		j = 0;
 	}
